feat(eskiz): bump eskizapi to spec with MessageItem.dispatch_id=string
The eskiz OpenAPI spec was wrong (dispatch_id typed as integer in
responses, but the live API returns it as a JSON string). Regenerated
client now models the field as NullableString; drop the ad-hoc direct
HTTP call we used to bypass the broken unmarshal.

+++ pkg/eskiz/service.go
@@ -155,7 +155,7 @@ func (s *service) SendBatch(ctx context.Context, messages []models.BatchMessage,
 	if o.DispatchID == 0 {
 		o.DispatchID = time.Now().UnixMilli()
 	}
-	dispatchID := float64(o.DispatchID)
+	dispatchID := int(o.DispatchID)
 
 	batchReq := eskizapi.SendSmsBatchRequest{Messages: inner, DispatchId: &dispatchID}
 	if s.cfg.CallbackURL() != "" {
